requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
pymysql
pydantic
email-validator
bcrypt
pytest
httpx